+++ internal/parser/source/source_test.go
@@ -353,6 +353,25 @@ up{node_status="v", job="node_exporter"}
 * on(instance) group_left(node_status) sliver_metadata
 `, // group_left on a label already guaranteed on the left
 	`services_enabled{job=""}`,
+	`
+group by (cluster, namespace, workload, workload_type, pod) (
+  label_join(
+    label_join(
+      group by (cluster, namespace, job_name, pod) (
+        label_join(
+          kube_pod_owner{job="kube-state-metrics", owner_kind="Job"}
+        , "job_name", "", "owner_name")
+      )
+      * on (cluster, namespace, job_name) group_left(owner_kind, owner_name)
+      group by (cluster, namespace, job_name, owner_kind, owner_name) (
+        kube_job_owner{job="kube-state-metrics", owner_kind!="Pod", owner_kind!=""}
+      )
+    , "workload", "", "owner_name")
+  , "workload_type", "", "owner_kind")
+)
+`,
+	`foo{job="xxx"} + on(job) group_right(instance) bar{}`,
+	`foo{job="xxx"} + ignoring(job) group_right(instance) bar{job="zzz"}`,
 }
 
 func TestLabelsSource(t *testing.T) {
@@ -373,6 +392,7 @@ func TestLabelsSource(t *testing.T) {
 			}
 			done[expr] = struct{}{}
 
+			t.Log(expr)
 			n, err := parser.DecodeExpr(expr)
 			if err != nil {
 				t.Error(err)